Fix import errors after splitting into files, add error checking for operation types that are missing tests
