+++ pxfish/pyfish.py
@@ -5,40 +5,14 @@ identified in the resources.py file.
 
 import os
 import argparse
-import re
-import json
 import logging
-import pull
-import push
-import paths
+from pull import *
+from push import *
+from paths import *
 from resources import resources
 from pydent import AqSession
 
 logging.basicConfig(level=logging.INFO)
-
-
-def field_type_list(field_types, role):
-    """
-    Returns the sublist of field types with the given role.
-
-    Arguments:
-      field_types (list): the list of field types
-      role (string): the role of field types to be returned
-
-    Returns:
-      list: the sublist of field_types that have the role
-    """
-    ft_list = []
-    for field_type in field_types:
-        if field_type.role == role:
-            ft_ser = {
-                "name": field_type.name,
-                "part": field_type.part,
-                "array": field_type.array,
-                "routing": field_type.routing
-            }
-            ft_list.append(ft_ser)
-    return ft_list
 
 def open_aquarium_session():
     """
